FileSystemFileStorage: Remove cache
Before this change, once a file has been loaded in the worker thread,
the ArrayBuffer was cached (with the file ID as lookup key). To be able
to use the file data in the frontend, we need to transfer the buffer to
the UI thread. After the transfer, the ArrayBuffer in the worker thread
is unusable. It is still part of the cache though, and will be returned
to the next caller that requests the same file ID. This is a problem.

In the long run, we'll have to cache files in the frontend thread
anyways, so having it in the SystemFileStore does not give any benefits.