Upgrade electron: 23.0.0 → 23.1.4
This upgrade addresses a bug on macOS that prevented Electron from
recognizing the file type (extension and Media Type) when a file was
pasted, or dragged and dropped, and its extension was present in the
file name but hidden in the Finder UI due to the 'Show all file
extensions' option being unchecked. This was a regression in electron
v23.0.0, which was fixed in v23.1.0 [1]. Now it works as expected again,
i.e. the full file name including extension is shown in the Media
Dialog, and the Media Type is properly inferred.

[1]: https://chromium.googlesource.com/chromium/src/+/9c3a77da9fa8c67b10