Add message length limit handling
The message length (in bytes) will be checked periodically (debounced on input) or before sending, and a text length indicator will be shown if it is close to (> 90%) or exceeding (> 100%) the limit of 1000 bytes. Message sending is blocked if the limit is exhausted. MediaFiles are also prevented from being sent if the caption is too long.

Amend implementation based on discussions.