Do not use 'count' as plural variable due to ICU format
When using i18next-parser, it creates all plural keys as expected by the
default i18next JSON format when it encounters the usage of the parameter
'count'. Since we are using ICU format instead, we have to disable that.
It is not currently possible to disable with i18next-parser , but as a
workaround, simple avoid naming that parameter 'count' and using e.g.
'number'.

See
https://github.com/i18next/i18next-parser/issues/463#issuecomment-1478570211