Implement logic to honour privacy settings to blocked contacts
This includes when a contact is explicitly blocked, i.e. whether its
identity string is present in the `blockedIdentities` list in the
privacy settings, as well as implicitly blocked, i.e. when the contact
is unknown and the `UnknownContactPolicy` option of the privacy settings
is set to `BLOCK_UNKNOWN`.

This commit also:
- removes an unused `Settings` class, which had been replaced by separated
  classes such as PrivacySettings`, `ProfileSettings`, etc.
- displays blocked icons in all necessary places, all reactive.
- replaces the message compose area with a warning message for a blocked
  contact to prevent the user from writing a message.
- limits the context menu actions for messages in a conversation with a
  blocked contact.
- hides blocked contacts from the contact list to forward a message.