Prevent electron-packager from downloading Electron checksum file
In @electron/get 2.0.3, a change was introduced to always download the
Electron release artifacts checksum file, to ensure that cached files
are actually untampered and complete. Unfortunately this causes issues
when building the app in a network-sandboxed environment.

Electron already ships the needed checksums, so we can just pass them to
@electron/get to avoid having the checksum file re-downloaded.

Additionally, we add a custom downloader which logs all electron
artifact downloads during the packaging phase, to simplify debugging in
the future.

This fixes the sandboxed Flatpak builds, which do not allow any network
requests during build time.