Revert usage of Popover API
Due to lack of support in our current version of TypeScript for the
new(-ish) Web Popover API, this commit reverts its usage back to the
previous layering strategy using z-index. This is currently not
problematic, as usage of top-layer features (introduced alongside
Popover) has been temporarily put on hold in the refactored Modal
component as well.