Remove --threema-persist-profile CLI flag
The non-persistent profile sits in a "temporary" directory, but when
closing and re-opening the app, the data is still there.

For this feature to be useful, we'd need to use an actual temporary
location. However, we don't currently make use of this flag, so it's
better to remove the code.