BlobCacheService: Fix logic of refreshCacheForMessage
The previous implementation was wrong: It did not update the store in
the cache, instead it replaced the store with another store. This means
that all existing subscribers did not get the refreshed store, but
instead a stale reference to the previous thumbnail store.

Additionally, we can add an optimization: If the store is not already
part of the cache, we don't need to do anything, because there are no
subscribers anyways.