Rename thumbnail generator / media service
The previous naming was a bit confusing: The `MediaService` wrapped a
`ThumbnailGenerator`, which was implemented by the
`FrontendMediaService`.

Additionally, the `ThumbnailGenerator` interface contained methods that
were not related to generating thumbnails, namely
`refreshCacheForMessage`.

To make it clearer where a certain service lives, and to make the name
more descriptive, the `ThumbnailGenerator` interface (which essentially
defines the `FrontendMediaService`) was renamed to
`IFrontendMediaService`. The `MediaService`, which wraps the remote
endpoint to the `IFrontendMediaService` was renamed to
`BackendMediaService`.