Only regenerate thumbnails after image download or upload
Previously, image thumbnails were _always_ regenerated when requesting
the image message blob. This is far from ideal. The code is now updated
to only regenerate the thumbnail after the main blob was downloaded from
the network. To achieve this, the `loadOrDownloadBlob` helper function
was updated to return the blob source.

Additionally, optimize the regenerating of thumbnails on upload: Return
the uploaded bytes from the `uploadBlobs` helper function to avoid
reloading (and redecrypting) the file bytes from the file system.