Regenerate thumbnail of outbound messages in task, not in `uploadBlobs`
This required a few changes to the message model controller:

- Return `UploadedBlobBytes` from `uploadBlobs()`
- Expose `uid` as public property

The `regenerateThumbnail` function can only be invoked from the model,
not from the task, because it needs access to the `db` service. Thus,
expose `regenerateThumbnail` through the outbound image message
controller.

I also optimized the required parameters passed to `regenerateThumbnail`.
With the new properties exposed on the message model controller, we can
simply pass `this` to `regenerateThumbnail` from the controller method.
This is cleaner than the previously required low-level parameters, like
`DbMessageUid` or the `ConverseationControllerHandle`.