Sign nightly macOS releases
Because notifications don't seem to work properly when a DMG is
unsigned, sign nightly releases as well.

Apple recommends not to notarize more than 75 apps per day. Since we
generate 5 releases per build for 2 architectures, that would be 7
merges to `develop` per day. That sounds acceptable, so let's try it.