Discard async RPC calls to the GC context if the connection was aborted
Rationale: A group call connection may be terminated by the main thread
in which case the abort event is dispatched asynchronously to the backend
worker where a call to `.sendP2s` or `.update` may still be in flight.
In this case, the pending RPC call is stale and we should ignore it.